# openapi/api-v1.yaml
openapi: 3.1.0
info:
  title: ILM Red API
  description: |
    Cloud-native API for digital knowledge management platform.

    ## Authentication

    The API supports two authentication methods:

    1. **Bearer Token (JWT)**: Include the JWT in the `Authorization` header
       ```
       Authorization: Bearer eyJhbGciOiJSUzI1NiI...
       ```

    2. **API Key**: Include your API key in the `X-API-Key` header
       ```
       X-API-Key: ilm_live_abc123...
       ```

    ## Rate Limiting

    | Tier       | Requests/min | AI Tokens/day |
    |------------|--------------|---------------|
    | Free       | 60           | 10,000        |
    | Premium    | 300          | 100,000       |
    | Enterprise | 1,000        | Unlimited     |

    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit`
    - `X-RateLimit-Remaining`
    - `X-RateLimit-Reset`

    ## Pagination

    List endpoints support pagination with `page` and `limit` query parameters.
    Response includes a `pagination` object with total count and page info.

  version: 1.0.0
  contact:
    name: ILM Red Support
    email: support@ilm-red.com
    url: https://docs.ilm-red.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.ilm-red.com/v1
    description: Production
  - url: https://api-staging.ilm-red.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /books:
    get:
      tags:
        - Books
      summary: List books
      operationId: listBooks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, private, friends]
        - name: category
          in: query
          schema:
            type: array
            items:
              $ref: '#/components/schemas/BookCategory'
        - name: language
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt, title, rating, views]
            default: createdAt
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of books
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Books
      summary: Upload a new book
      operationId: createBook
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - title
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF, EPUB, or TXT file (max 500MB)
                title:
                  type: string
                  maxLength: 500
                author:
                  type: string
                  maxLength: 200
                description:
                  type: string
                  maxLength: 5000
                category:
                  $ref: '#/components/schemas/BookCategory'
                visibility:
                  type: string
                  enum: [public, private, friends]
                  default: private
                language:
                  type: string
                  description: ISO 639-1 language code
      responses:
        '201':
          description: Book created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          $ref: '#/components/responses/ValidationError'
        '413':
          description: File too large
...
  /books/{bookId}/pages:
    get:
      tags:
        - Books
      summary: List book pages with image URLs
      description: Get paginated list of all pages with signed URLs for each resolution
      operationId: listBookPages
      parameters:
        - $ref: '#/components/parameters/BookIdParam'
        - name: resolution
          in: query
          description: Filter by specific resolution
          schema:
            type: string
            enum: [thumbnail, medium, high-res, ultra]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Page manifest with URLs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageManifest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /books/{bookId}/pages/generate:
    post:
      tags:
        - Books
      summary: Trigger page image generation
      description: Start background job to generate page images for a book
      operationId: generateBookPages
      parameters:
        - $ref: '#/components/parameters/BookIdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                force:
                  type: boolean
                  default: false
                  description: Regenerate even if pages already exist
                resolutions:
                  type: array
                  items:
                    type: string
                    enum: [thumbnail, medium, high-res, ultra, all]
                  default: [all]
                  description: Resolutions to generate
                pageRange:
                  type: object
                  description: Optional page range to generate
                  properties:
                    start:
                      type: integer
                      minimum: 1
                    end:
                      type: integer
                      minimum: 1
      responses:
        '202':
          description: Generation job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageGenerationJob'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Generation already in progress
...
  /chats:
    get:
      tags:
        - Chat
      summary: List user's chat sessions
      description: Get paginated list of chat sessions for the authenticated user.
      operationId: listChatSessions
      parameters:
        - name: archived
          in: query
          description: Filter by archived status
          schema:
            type: boolean
        - name: bookId
          in: query
          description: Filter by book ID
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of chat sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatSession'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Chat
      summary: Create new chat session
      description: |
        Create a new chat session. Sessions can be standalone or linked to a book
        for context-aware AI chat.
      operationId: createChatSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatSessionCreate'
      responses:
        '201':
          description: Chat session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSession'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /chats/{sessionId}/messages:
    get:
      tags:
        - Chat
      summary: Get messages in a chat session
      description: Get paginated messages for a chat session.
      operationId: getChatMessages
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: order
          in: query
          description: Sort order (asc = oldest first, desc = newest first)
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: Chat messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatMessage'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags:
        - Chat
      summary: Send message (non-streaming)
      description: |
        Send a message to the AI and receive a complete response.
        For real-time streaming, use POST /chats/{sessionId}/stream instead.
      operationId: sendChatMessage
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageCreate'
      responses:
        '200':
          description: AI response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
        '402':
          description: Insufficient credits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InsufficientCreditsError'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /chats/{sessionId}/stream:
    post:
      tags:
        - Chat
      summary: Send message with streaming response (SSE)
      description: |
        Send a message to the AI and receive a streaming response via Server-Sent Events.

        **Event Types:**
        - `chunk`: Content chunk `{"type": "chunk", "content": "Hello"}`
        - `tokens`: Token count update `{"type": "tokens", "input": 50, "output": 10}`
        - `done`: Stream complete `{"type": "done", "finish_reason": "stop", "message_id": "uuid", "cost_cents": 2}`
        - `error`: Error occurred `{"type": "error", "code": "RATE_LIMITED", "message": "..."}`

        **Client Implementation:**
        Use EventSource API (web) or SSE library (mobile) to consume the stream.
      operationId: streamChatMessage
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageCreate'
      responses:
        '200':
          description: SSE stream
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
        '402':
          description: Insufficient credits
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users
      operationId: adminListUsers
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, deleted]
        - name: role
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUser'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{userId}:
    patch:
      tags:
        - Admin
      summary: Update user
      operationId: adminUpdateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, suspended]
                roles:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: User updated
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
...